2.10.0

+++ Carbon/Functions/Install-Certificate.ps1
@@ -6,24 +6,32 @@ function Install-CCertificate
     Installs a certificate in a given store.
     
     .DESCRIPTION
-    Uses the .NET certificates API to add a certificate to a store for the machine or current user.  The user performing the action must have permission to modify the store or the installation will fail.
+    Uses the .NET certificates API to add a certificate to a store for the machine or current user.  The user performing
+    the action must have permission to modify the store or the installation will fail.
 
-    To install a certificate on a remote computer, create a remoting session with the `New-PSSession` cmdlet, and pass the session object to this function's `Session` parameter. When installing to a remote computer, the certificate's binary data is converted to a base-64 encoded string and sent to the remote computer, where it is converted back into a certificate. If installing a certificate from a file, the file's bytes are converted to base-64, sent to the remote computer, saved as a temporary file, installed, and the temporary file is removed.
+    To install a certificate on a remote computer, create a remoting session with the `New-PSSession` cmdlet, and pass
+    the session object to this function's `Session` parameter. When installing to a remote computer, the certificate's
+    binary data is converted to a base-64 encoded string and sent to the remote computer, where it is converted back
+    into a certificate. If installing a certificate from a file, the file's bytes are converted to base-64, sent to the
+    remote computer, saved as a temporary file, installed, and the temporary file is removed.
 
     The ability to install a certificate on a remote computer was added in Carbon 2.1.0.
     
     .OUTPUTS
-    System.Security.Cryptography.X509Certificates.X509Certificate2. An X509Certificate2 object representing the newly installed certificate.
+    System.Security.Cryptography.X509Certificates.X509Certificate2. An X509Certificate2 object representing the newly
+    installed certificate.
     
     .EXAMPLE
     > Install-CCertificate -Path C:\Users\me\certificate.cer -StoreLocation LocalMachine -StoreName My -Exportable -Password My5up3r53cur3P@55w0rd
     
-    Installs the certificate (which is protected by a password) at C:\Users\me\certificate.cer into the local machine's Personal store.  The certificate is marked exportable.
+    Installs the certificate (which is protected by a password) at C:\Users\me\certificate.cer into the local machine's
+    Personal store.  The certificate is marked exportable.
     
     .EXAMPLE
     Install-CCertificate -Path C:\Users\me\certificate.cer -StoreLocation LocalMachine -StoreName My -ComputerName remote1,remote2
     
-    Demonstrates how to install a certificate from a file on the local computer into the local machine's personal store on two remote cmoputers, remote1 and remote2. Use the `Credential` parameter to connect as a specific principal.
+    Demonstrates how to install a certificate from a file on the local computer into the local machine's personal store
+    on two remote cmoputers, remote1 and remote2. Use the `Credential` parameter to connect as a specific principal.
     #>
     [CmdletBinding(SupportsShouldProcess=$true,DefaultParameterSetName='FromFileInWindowsStore')]
     [OutputType([Security.Cryptography.X509Certificates.X509Certificate2])]
@@ -76,22 +84,34 @@ function Install-CCertificate
         # Use the `Session` parameter to install a certificate on remote computer(s) using PowerShell remoting. Use `New-PSSession` to create a session.
         #
         # This parameter was added in Carbon 2.1.0.
-        $Session
+        $Session,
+
+        # Re-install the certificate, even if it is already installed. Calls the `Add()` method for store even if the
+        # certificate is in the store. This function assumes that the `Add()` method replaces existing certificates.
+        [switch]$Force,
+
+        [switch]$NoWarn
     )
     
     Set-StrictMode -Version 'Latest'
-
     Use-CallerPreference -Cmdlet $PSCmdlet -Session $ExecutionContext.SessionState
+
+    if( -not $NoWarn )
+    {
+        $msg = 'Carbon''s "Install-CCertificate" function is OBSOLETE and will be removed in the next major version ' +
+               'of Carbon. Use the "Install-CCertificate" function in the new "Carbon.Cryptography" module.'
+        Write-CWarningOnce -Message $msg
+    }
 
     if( $Password -and $Password -isnot [securestring] )
     {
-        Write-Warning -Message ('You passed a plain text password to `Install-CCertificate`. A future version of Carbon will remove support for plain-text passwords. Please pass a `SecureString` instead.')
+        Write-CWarningOnce -Message ('You passed a plain text password to `Install-CCertificate`. A future version of Carbon will remove support for plain-text passwords. Please pass a `SecureString` instead.')
         $Password = ConvertTo-SecureString -String $Password -AsPlainText -Force
     }
     
     if( $PSCmdlet.ParameterSetName -like 'FromFile*' )
     {   
-        $resolvedPath = Resolve-Path -Path $Path        
+        $resolvedPath = Resolve-Path -Path $Path
         if( -not $resolvedPath )
         {
             return
@@ -102,26 +122,31 @@ function Install-CCertificate
         $fileBytes = [IO.File]::ReadAllBytes($Path)
         $encodedCert = [Convert]::ToBase64String( $fileBytes )
 
-        $keyFlags = [Security.Cryptography.X509Certificates.X509KeyStorageFlags]::MachineKeySet
+        # Make sure loading the certificate doesn't leave temporary cruft around on the file system. We're only loading
+        # the cert to get its thumbprint.
+        $keyStorageFlags = @{}
         if( $StoreLocation -eq [Security.Cryptography.X509Certificates.StoreLocation]::CurrentUser )
         {
-            $keyFlags = [Security.Cryptography.X509Certificates.X509KeyStorageFlags]::UserKeySet
+            $keyStorageFlags['KeyStorageFlags'] = 
+                [Security.Cryptography.X509Certificates.X509KeyStorageFlags]::EphemeralKeySet
         }
-        $keyFlags = $keyFlags -bor [Security.Cryptography.X509Certificates.X509KeyStorageFlags]::PersistKeySet
-    
-        if( $Exportable )
-        {
-            $keyFlags = $keyFlags -bor [Security.Cryptography.X509Certificates.X509KeyStorageFlags]::Exportable
-        }
-
-        $Certificate = Get-CCertificate -Path $Path -Password $Password -KeyStorageFlags $keyFlags
-        $fromFile = $true
+        $Certificate = Get-CCertificate -Path $Path -Password $Password -NoWarn @keyStorageFlags
     }
     else
     {
         $encodedCert = [Convert]::ToBase64String( $Certificate.RawData )
-        $keyFlags = 0
-        $fromFile = $false
+    }
+
+    $keyFlags = [Security.Cryptography.X509Certificates.X509KeyStorageFlags]::MachineKeySet
+    if( $StoreLocation -eq [Security.Cryptography.X509Certificates.StoreLocation]::CurrentUser )
+    {
+        $keyFlags = [Security.Cryptography.X509Certificates.X509KeyStorageFlags]::UserKeySet
+    }
+    $keyFlags = $keyFlags -bor [Security.Cryptography.X509Certificates.X509KeyStorageFlags]::PersistKeySet
+
+    if( $Exportable )
+    {
+        $keyFlags = $keyFlags -bor [Security.Cryptography.X509Certificates.X509KeyStorageFlags]::Exportable
     }
 
     $invokeCommandArgs = @{ }
@@ -130,36 +155,40 @@ function Install-CCertificate
         $invokeCommandArgs['Session'] = $Session
     }
 
-
     Invoke-Command @invokeCommandArgs -ScriptBlock {
         [CmdletBinding()]
         param(
-            [Parameter(Mandatory=$true)]
-            [string]
+            [Parameter(Mandatory,Position=0)]
             # The base-64 encoded certificate to install.
-            $EncodedCertificate,
+            [String]$EncodedCertificate,
 
             # The password for the certificate.
-            [securestring]
-            $Password,
+            [Parameter(Position=1)]
+            [securestring]$Password,
 
-            [Parameter(Mandatory=$true)]
-            [Security.Cryptography.X509Certificates.StoreLocation]
-            $StoreLocation,
+            [Parameter(Mandatory,Position=2)]
+            [Security.Cryptography.X509Certificates.StoreLocation]$StoreLocation,
         
+            [Parameter(Position=3)]
             $StoreName,
 
-            [string]
-            $CustomStoreName,
+            [Parameter(Position=4)]
+            [String]$CustomStoreName,
 
-            [Security.Cryptography.X509Certificates.X509KeyStorageFlags]
-            $KeyStorageFlags,
+            [Parameter(Position=5)]
+            [Security.Cryptography.X509Certificates.X509KeyStorageFlags]$KeyStorageFlags,
 
-            [bool]
-            $WhatIf,
+            [Parameter(Position=6)]
+            [bool]$WhatIf,
 
-            [Management.Automation.ActionPreference]
-            $Verbosity
+            [Parameter(Position=7)]
+            [Management.Automation.ActionPreference]$Verbosity,
+
+            [Parameter(Position=8)]
+            [switch]$Force,
+
+            [Parameter(Mandatory,Position=9)]
+            [String]$Thumbprint
         )
 
         Set-StrictMode -Version 'Latest'
@@ -169,16 +198,9 @@ function Install-CCertificate
 
         $tempDir = 'Carbon+Install-CCertificate+{0}' -f [IO.Path]::GetRandomFileName()
         $tempDir = Join-Path -Path $env:TEMP -ChildPath $tempDir
-        New-Item -Path $tempDir -ItemType 'Directory' -WhatIf:$false | Out-Null
 
         try
         {
-            $certBytes = [Convert]::FromBase64String( $EncodedCertificate )
-            $certFilePath = Join-Path -Path $tempDir -ChildPath ([IO.Path]::GetRandomFileName())
-            [IO.File]::WriteAllBytes( $certFilePath, $certBytes )
-
-            $cert = New-Object 'Security.Cryptography.X509Certificates.X509Certificate2' ($certFilePath, $Password, $KeyStorageFlags)
-
             if( $CustomStoreName )
             {
                 $store = New-Object 'Security.Cryptography.X509Certificates.X509Store' $CustomStoreName,$StoreLocation
@@ -188,6 +210,30 @@ function Install-CCertificate
                 $store = New-Object 'Security.Cryptography.X509Certificates.X509Store'  ([Security.Cryptography.X509Certificates.StoreName]$StoreName),$StoreLocation
             }
 
+            if( -not $Force )
+            {
+                $store.Open( ([Security.Cryptography.X509Certificates.OpenFlags]::ReadOnly) )
+                try
+                {
+                    if( $store.Certificates | Where-Object { $_.Thumbprint -eq $Thumbprint } )
+                    {
+                        return
+                    }
+                }
+                finally
+                {
+                    $store.Close()
+                }
+            }
+
+            # Only do the file system work if the certificate isn't installed.
+            New-Item -Path $tempDir -ItemType 'Directory' -WhatIf:$false | Out-Null
+
+            $certBytes = [Convert]::FromBase64String( $EncodedCertificate )
+            $certFilePath = Join-Path -Path $tempDir -ChildPath ([IO.Path]::GetRandomFileName())
+            [IO.File]::WriteAllBytes( $certFilePath, $certBytes )
+
+            $cert = New-Object 'Security.Cryptography.X509Certificates.X509Certificate2' ($certFilePath, $Password, $KeyStorageFlags)
             $store.Open( ([Security.Cryptography.X509Certificates.OpenFlags]::ReadWrite) )
 
             $description = $cert.FriendlyName
@@ -205,10 +251,13 @@ function Install-CCertificate
         }
         finally
         {
-            Remove-Item -Path $tempDir -Recurse -ErrorAction Ignore -WhatIf:$false -Force
+            if( (Test-Path -Path $tempDir) )
+            {
+                Remove-Item -Path $tempDir -Recurse -ErrorAction Ignore -WhatIf:$false -Force
+            }
         }
 
-    } -ArgumentList $encodedCert,$Password,$StoreLocation,$StoreName,$CustomStoreName,$keyFlags,$WhatIfPreference,$VerbosePreference
+    } -ArgumentList $encodedCert,$Password,$StoreLocation,$StoreName,$CustomStoreName,$keyFlags,$WhatIfPreference,$VerbosePreference,$Force,$Certificate.Thumbprint
 
     return $Certificate
 }
